+++ apis/Authentication/hashHelpers.js
